requirements.txt
customtkinter
pillow
ultralytics
pyyaml
opencv-python
